Modules/UdosLocalLlm/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "udos-local-llm"
version = "0.1.0"
description = "uDos Round 8 — local knowledge library, Chroma RAG, Ollama integration (privacy-first)"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "typer>=0.12.0",
  "httpx>=0.27.0",
  "feedparser>=6.0.11",
  "pyyaml>=6.0.1",
  "chromadb>=0.5.0",
  "beautifulsoup4>=4.12.0",
  "html2text>=2024.2.26",
  "pypdf>=5.0.0",
]

[project.optional-dependencies]
dev = ["pytest>=8.0.0", "ruff>=0.6.0"]

[project.scripts]
udos-llm = "udos_llm.cli:run"

[tool.hatch.build.targets.wheel]
packages = ["udos_llm"]

[tool.ruff]
line-length = 100
target-version = "py311"
